dynamicli/requirements.txt
dnspython
python-whois
scapy
validators
requests
bs4
dynacli
pyfiglet